requirements.txt
Django
Markdown
Pygments
pytz
sqlparse
whoosh
django-haystack
jieba
django-mdeditor